Navigation: Fixed nav bar bugs. Changed post-login action
Fixed bug where navigation bar was displaying the wrong background image.
Changed functionality so that after logging in, the user will
have to manually re-choose his/her requested action. For example, if
the user wishes to upload a picture, but he is not logged in the app
will go to the login screen. After logging in, the user will have
to hit "back", where the root view controller will be loaded.
Once at the root view, the user will have to choose "upload" again.