TextFields: Save text as it is entered
Found a bug where form is submitted (login/upload) when a textfield
has focus, but is not in view because of scrolling, and button is
pressed. What happens is that the text never gets saved in an array.
The form submission should validate text in the array, not text
directly from the textfield. This is why it is important to save
text in the array as the user enters data.