Login: New create account trigger
Change trigger from table view cell to label. The initial layout
was confusing because it appeared as if user had to create an
account as a pre-requisite to logging in. Moving "create" so that
it's underneath the bottom makes it more clear that creating an
account is only optional (ie. if user doesn't already have acct).